Still not working but went to Vinix's style of oneshot (Finding the frequency of the clock then using that to schedule)

+++ src/include/drivers/pit.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 void sleep(uint64_t ticks);
-uint32_t pit_read_count();
+uint16_t pit_read_count();
 int init_pit();
 
 #endif
